refactor(schema,nuxt): move `NuxtLinkOptions` into `@nuxt/schema`
